use the same image for control plane and message queue
