--- docker/base.py
@@ -1,4 +1,0 @@
-from importlib.metadata import version
-
-
-print(version("llama-deploy"))
